demos/redbank-demo/knowledge-agent/pyproject.toml
[project]
name = "redbank-knowledge-agent"
version = "1.0.0"
description = "Read-only knowledge and data retrieval agent (A2A) for the RedBank demo — RAG and customer data via MCP."
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]==0.3.26",
    "langchain>=1.2.15",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=1.1.13",
    "langgraph>=1.1.6",
    "mlflow[kubernetes]>=3.11.1",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=1.0.0",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
